--separate managed container functions in core
