+++ src/esp/bindings/CoreBindings.cpp
@@ -4,7 +4,6 @@
 
 #include "esp/bindings/bindings.h"
 #include "esp/core/Configuration.h"
-#include "esp/core/ManagedContainer.h"
 #include "esp/core/random.h"
 
 namespace py = pybind11;
