+++ src/esp/core/managedContainers/AbstractManagedObject.h
@@ -5,7 +5,7 @@
 #ifndef ESP_CORE_ABSTRACTMANAGEDOBJECT_H_
 #define ESP_CORE_ABSTRACTMANAGEDOBJECT_H_
 
-#include "esp.h"
+#include "esp/core/esp.h"
 
 namespace esp {
 namespace core {
